+++ src/XSRenderer/XSFramebuffer.cpp
@@ -114,7 +114,7 @@ namespace XS {
 			glDeleteFramebuffersEXT( 1, &id );
 		}
 
-		void Framebuffer::AttachColorTexture( const Texture *texture, size_t slot ) {
+		void Framebuffer::AttachColorTexture( const Texture *texture, unsigned int slot ) {
 			if ( slot >= MAX_FBO_COLOR_TEXTURES ) {
 				Console::Print( "Invalid slot number given (%d), valid range is 0 - %d", slot, MAX_FBO_COLOR_TEXTURES-1 );
 				return;
